week-1/requirements.txt
ptyhon-dotenv
langchian
langchain_google_genai
python-json-logger
yaml